requirements.txt
google-cloud-storage==2.15.0
google-cloud-bigquery==3.19.0
scikit-learn==1.4.1.post1
dask==2024.2.1
pandas==2.2.1
numpy==1.26.4
pandas-gbq==0.22.0
PyYAML==6.0.1
kaleido==0.2.1.post1